Handle runtime error spans from different packages (#675)
Fixes #654.

Evaluator errors contain source spans. These source spans sometimes come
from a package that is different than the one being currently executed.
Previously, to report errors with source code context, we would always
use the `SourceMap` associated with the currently executing package.
Now, to track which package the source span actually belongs to, we use
a new `PackageSpan` type (package ID / span pair) instead of a `Span`
type.

When we later try to resolve the source code for reporting, we look up
the `PackageId` from the error in the package store to locate the
correct `SourceMap`.

I also had to change in the interpreter to make the lookup work:
Previously, we weren't actually storing the _current_ package (the "user
package" as it's sometimes referred to in the code) in the package
store. So the `PackageId` lookup would succeed for std/core libs, but
fail for the current package! With these changes, we now insert the
`CompileUnit` that backs the current package into the package store. As
we append sources to the compilation, we update the `CompileUnit` in the
store with the new sources.

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>