Refresh the tree view on new job submission and pending jobs (#716)
Fixes #705 

Whenever a new job is submitted, or a workspace is added to the tree
that has not-yet-completed jobs, the workspace will refresh in the tree
for a period using a back-off latency up to a maximum time period. It
will stop refreshing a workspace once the submitted jobs are present and
all jobs are in a completed state in the tree.

If the user manually refreshes the tree via the refresh icon, the
process outlined above will start again.