Update error handling of floating doc comments (#688)
Closes #652 

Checks if doc comments on an `Item` that fails to parse are non-empty.
Pushes an error, `ErrorKind::FloatingDocComment(span)`, and returns an
error `Item` so that the parser will continue. This returns only a
single, more helpful error message for floating doc comments and doesn't
consume the closing curly brace.

Notes: If compilation _shouldn't fail_ because of floating doc comments,
there could be a recovery step for the error that also removes the error
`Item`.

---------

Co-authored-by: Chris O'Brien <obrchr96@gmail.com>
Co-authored-by: Alex Hansen <alex@alex-hansen.com>
Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>