Fix cross-unit UDT dependency (#766)
This fixes a subtle bug in usage of UDTs across compilation units.
Because a UDT is captured as a type that includes a `Res` in it, it wil
have the placeholder `None` for it's package id when initially compiled
(package ids are only known after insertion into the store). Later, a
type mismatch could occur because the caller has the package id in the
UDT type but the signature stored in the scheme uses UDT types with
`None`. This fix adds a notion of updating a scheme to include the
package id at type checking time by using `with_package` such that type
checking for both internal/`None` resolutions and dependencies from
other compilation units with known package ids can all be properly used.