Fix wasm panic at 'ExprId should have been lowered' (#781)
When executing the entry expression for a compilation in the stateful
interpreter, it should be treated the same way as running shots for
Python or executing for code generation: use a fresh execution state and
environment. State reuse is only meaningful in steppable debugging, and
environment reuse is only for fragment execution in the contex of a REPL
or Python/Jupyter. The fix used here is very small (just the change in
stateful.rs), the rest is introducing test infra and cases for the wasm
crate.

Fixes #777