Update to Rust 1.73, improve devcontainer
This change updates the build scripts and yaml to use Rust 1.73, which introduced some new clippy lints that the code was in violation of. The changes have the version update and the code updates for compatibility, and to keep the new clippy lints in check the minimum required version is updated in prereqs.py.

This also updates the devcontainer definition to include the Azure CLI and fixes the Python version to 3.11 (3.12 was releast on Oct 3, 2023 and does not yet play nice with azure-cli packages, see https://github.com/microsoft/azure-devops-python-api/pull/469).