Update to Rust 1.73 (#784)
This change updates the build scripts and yaml to use Rust 1.73, which
introduced some new clippy lints that the code was in violation of. The
changes have the version update and the code updates for compatibility,
and to keep the new clippy lints in check the minimum required version
is updated in prereqs.py.

This also updates the devcontainer to pin Python 3.11 because latest is
now pointing to 3.12 which fails to build during container creation.