Signature Help for Indirect Calls Based on Type (#791)
Adds signature help to the language server for indirect call expressions
based on the type of the callee expression. This allows signature help
to be given for all call expressions, including lambdas, local captures
of callables and partial applications, functor applications, arrays of
callables, and callables returned from a block, and UDT constructors.