Introduce Q# project structure (#794)
# Q# Projects

This PR introduces a Q# manifest file, `qsharp.json`, and a
corresponding notion of _Q# projects_.

## Feature Overview
When the compiler is invoked, it looks upward in the directory structure
for a manifest file. Upon finding a valid manifest, we set that
directory as the root directory of the project. We then include all
`*.qs` files in that directory or subdirectories of
that directory as sources, unless they are explicitly excluded in the
manifest. This implementation *does not follow symlinks of any kind*.

All FS requirements are expressed as the `FileSystem` interface. This
will provide compatibility when we integrate projects with the language
service and other virtual file systems. `qsc_project` only utilizes
`std::fs` if the `fs` feature is enabled.

## API Changes

There are some notable differences in the behavior of `qsc` and `qsi`:
* `qsc` still accepts sources as input, but now, if no sources are
specified, `qsc` will look for a directory in the current directory.
* Currently, `qsc` requires a list of source names on the command line.
With this PR, if `qsc` is invoked within a project with
  no sources specified, it will include project sources.
* `qsi` will include sources within its project scope if it is invoked
within a project. A flag `--no-init-project`/`-n`
   has been added to skip this.

## Testing
`qsc_project` is tested using our normal `expect_test` workflow. There
are a series of test projects in `qsc_project/tests` that represent
various project structures and manifests. We expect certain directory
structures and manifests to deterministically correspond to a set of
source names and source contents, and this is the interface that is
tested.