In memory file system (#801)
If we want to enable users to click a link and land on vscode.dev with
code passed in the URL, we'll need some type of in memory file system
for it. This is an outline for that. For now, I've wired it in with a
simple command to open all our sample algorithms in a kind of sandbox
(so you can play around with them, change, debug, etc. all without being
backed by files on a disk or a repo. Any changes are lost when you close
it).

It does look like you're expected to vendor in your own in-memory file
system implementation, even though this is a common feature. (I'm
discussing with the VS Code team to confirm, and have the question at
https://stackoverflow.com/questions/77312576/in-memory-file-system-for-vs-code-extension
). I'll update if there's a simpler approach.