Fix handling of stdlib UDTs in language service (#814)
Fixes #808.

Basically, when follow a reference to a UDT in the HIR through
[`qsc_hir::ty::Ty::Udt(res)`](https://github.com/microsoft/qsharp/blob/70b107ca8bb7575bb5c7861b17788e217d12f71d/compiler/qsc_hir/src/ty.rs#L42C14-L42C14),
we need to be cognizant of the `PackageId` that that UDT reference came
from.

Why?

`qsc_hir::hir::ItemId` uses `None` package id to indicate "local"
package.
https://github.com/microsoft/qsharp/blob/70b107ca8bb7575bb5c7861b17788e217d12f71d/compiler/qsc_hir/src/hir.rs#L172-L177

`qsc_hir::hir::Res` uses `ItemId` to associate HIR nodes with
resolutions.


https://github.com/microsoft/qsharp/blob/70b107ca8bb7575bb5c7861b17788e217d12f71d/compiler/qsc_hir/src/hir.rs#L191-L198

When the resolution lives in the same package as the HIR node, the
`PackageId` in `ItemId` is `None`, to indicate the "local" package. The
problem is, the "local" package is relative to where the HIR node came
from in the first place. So if all we have is the HIR node struct, we
don't have the ability to resolve the item correctly in the
`PackageStore`.


The fix is to make sure we keep track of the current `PackageId` when
we're following UDT resolutions, and when we're passing around any HIR
node that _could_ contain a UDT reference. This involves a lot of
annoying plumbing of the ID through all of the language service
features.