Un-deprecate `checkCode` (#823)
Apparently, the website intends to use this to check the validity of
generated Q# snippets. This seems like a legitimate use case. It's a
perfectly good API :)

My original intention for deprecating `checkCode` was to guide people
towards using the language service for editor squiggles, and that
mission has already been accomplished...