Add ability to continue on error for incremental compiler (#820)
This change is required for the notebook language service (#759). The
new public method is not used in this PR yet. It will be used in #759.

This change adds two flavors of `compile_fragments` to the incremental
compiler: one that fails fast and one that accumulates errors but
continues compilation.

When calling from the interpreter, we want to fail fast when parsing
fails, so that we don't pollute the resolver/checker/lowerer state with
flawed data when we already _know_ that cell execution will fail.

On the other hand, when calling from the language service, we want to
continue compilation even in the case of errors, so we can finish
building our AST & HIR we will use for editor features. Also, this way
we can report resolve/check errors even if there are syntax error
elsewhere in the program (e.g. in another cell).