Support Unimplemented attribute
This change introduces the @Unimplemented() attributed which can be used to mark an item as something that should not be used. The motivation behind including this and not just treating it as a Not Found error is to give the user more information that this API may have previously existing and is not currently supported (see #699 for context).