Disable doc tests (#829)
We have three instances of doc tests that right now take an unusually
long time for little benefit. For now, this change modifies the
Cargo.toml settings in the repo to disable doc tests and avoid the extra
cost during `cargo test --release`.

In the CI pipeline, the tests would take around 11 seconds each, due to
"fat" LTO. With them disabled that drops to zero since they are skipped
entirely.