Use `Default` to handle missing entry in sequence
This uses a more complex but ultimately more correct approach for fixing #672. By creating default error kinds for several of the AST and HIR types that can appear in sequences, sequence parsing recovery now inserts error entries to correspond to places where commas appear at the start of a sequence or back-to-back in the sequence. The end result is better signature help due to recovered HIR including placeholder error entries.