Allow i64::MIN literals of any radix
Previously, although the case of literals overflowing to i64::MIN was
treated specially in #635, it was only checked if the overflow occurred
during the addition step.  This is fine for decimal literals, but in the
case of hexadecimal literals, the overflow occurs when multiplying the
most significant digit by the place value, and in the case of binary
literals, the overflow occurs when advancing the place value.

This commit fixes hexadecimal and binary i64::MIN literals so they no
longer cause an error, but it is still an error if there is overflow
beyond i64::MIN (as intended).