Allow reuse of unmeasured qubits in QIR codegen (#843)
The previous change to disallow qubit reuse was too aggresive with how
it used fresh qubit ids, causing programs to get too big too fast. The
problem was the assumption that a qubit that is released can never be
used again. Instead, this updates the logic to ensure that qubits that
are released without ever having been measured or reset can be reused in
later allocations. As an example, this takes the Grover's Search sample
program from using 29 qubits to using 9 by allowing reuse of unmeasing
auxiliary qubits used during decomposition of multi-controlled gates.