Use `Default` to handle missing entry in sequence (#836)
This uses a more complex but ultimately more correct approach that fixes
#672. By creating default error kinds for several of the AST and HIR
types that can appear in sequences, sequence parsing recovery now
inserts error entries to correspond to places where commas appear at the
start of a sequence or back-to-back in the sequence. The end result is
better signature help due to recovered HIR including placeholder error
entries.