Fuzz: "thread '<unnamed>' panicked at compiler/qsc_frontend/src/typeck/rules.rs:368:57:" (ubuntu-latest)
When an attriute expression includes the nested definition of an item, we were incorrectly checking the type signatures of that item even though it was not included in early parts of type checking, leading to a panic. This update correctly ignores items in attributes for both phases of type checking instead of just the first one.

Fixes #846