Add `dump_machine` to Python API (#853)
This adds a `dump_machine` API that allows Python code to get the
allocated qubit count and dictionary of sparse state amplitudes in
Python. This will enable more test scenarios in Python for Q# code.

This also includes an update to how `BigInt` values are returned in
Python by using the arbitrary bitwidth Python integer type instead of
formatting them as a string.


![image](https://github.com/microsoft/qsharp/assets/10567287/44a4c014-dfaa-41c6-b0ee-b18de8a5d39c)