Use the storage proxy to avoid CORS issues (#866)
Services added a proxy API for storage requests to avoid the CORS
issues. It closely maps the (undocumented) storage proxy endpoint &
config we already supported, so I'm making the minimal changes here to
use that by default. Once we're happy it's stable, I'll take out the old
code and just leave the proxy path in.