Refactor UX components and add WebViews to VS Code (#865)
There is more I want to do, but this is big enough already, and want to
get the major pieces in.

This refactors some of the Preact UX components into the NPM package and
then shares them between the playground and VS Code.

It adds histograms, re, and a help page rendering as WebViews in VS
Code.

The help page (content still to be fleshed out) will open to the side
when the Help command is run, and this will be done by default when
opening a code sharing link that takes you to vscode.dev/quantum.