LS and Error Display Refactor (#862)
Refactors the language server and error display of type parameters and
UDT's. These types now have `Rc<str>` for their names, avoiding us to
drop a lot of complex lookup logic. The display logic for the types was
also moved from the language server's display.rs to the `.display()` on
the `Ty` struct.
As a result of these changes, error messages can now print a
user-friendly display of types like type parameters and UDT's.