Project Support in the VS Code Extension (#849)
# Project Support for the Language Service

This PR adds multi-file projects to the language service. Closes #796 

## Overview

`qsc_project` defines a project system that sits on top of a given file
system-like interface. This PR is an implementation of that interface
for the language service. This is a non-trivial implementation, as it
has cascading effects. The most prominent of these effects is the
asyncification of the language service, both in WASM and in TS.

## Async Information

There is some key behavior related to the asynchronous functions that
needs to be highlighted.
- We only trigger one compilation at a time, storing an optional "next"
compilation to catch up to the editor state. This prevents race
conditions when we fire off a compile event every keystroke.  
- The WASM now uses `async`/`await` in Rust. It is important to know
that these Rust `Future`s are equivalent to Javscript `Promise`s. They
share the same event loop scheduler. `wasm_bindgen` handles the sharing
of the async runtime between v8 and Rust, and it is indeed magical.
- `tokio` was added to the workspace dependencies. This is a bummer, but
it is only used for running the unit tests in the language service. We
do not use tokio's runtime anywhere in our non-test code.
- I've done my best to architect this in such a way that it is easy to
debug. That being said, there are some rough edges. I will schedule some
time to go over those rough edges.

## Debugger
This PR does not introduce debugger support for projects. We will add
debugger support in a future PR. For now, this PR disallows running the
debugger if the current context is a Q# project. See #797.

![image](https://github.com/microsoft/qsharp/assets/12157751/c5ef17c9-f5a5-4e4e-8176-1c5f6d56f5b2)

## Follow-On Tasks
Below are the follow-on tasks that will come in future PRs:
- Using VS Code's FS watcher API to notify us of file updates, and
caching files.
- Redesigning the manifest format and `excludeFiles`/`excludeRegexes`
- Adding project support to the Python API (#861)

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>