Update UI to use profiles, compiler to use capabilities. (#888)
The capabilities are pulled from Cesar's spike PR trying to lay the
integration point so that we have the compiler only ever works with
capabilities. The profiles are currently defined in `qsc` so that the
compiler can't take a dependency on profiles.

I'm not sure how many capabilities we'll need to map, so the flags enum
may need to change to `u64` or a bit array.

Still not sure what to call the values in the `@Config()` attribute.
Currently `Base`/`Full` are used. `None`/`Any` has been proposed.

Effects #863 
Replaces #842