Bring back red squigglies in playground (update file version when in SFM) (#898)
This error was caused by a document version mismatch when reporting
diagnostics back to the
playground. 

We perform a manual version check
[here](https://github.com/microsoft/qsharp/blob/main/playground/src/editor.tsx#L106),
which was failing. While
removing that version check would have restored diagnostic reporting, it
would have
masked the underlying bug. Even though we don't use versions for
anything, this version
mismatch would have gone undetected for a while and perhaps caused other
issues. For that reason,
I am happy we have the above linked forward-thinking version check! 

Ultimately, the source of the version mismatch was failing to update
`open_documents` in the language
service when running in single file mode.