Use more targeted spans for type equality errors
This updates a few spots where type equality constraints were used in the type checker and had the whole expression span rather than just a component of the expression. This makes it easier to identify specific parts of an expression with type errors, especially for larger expressions like addition of multiple variables.