Multifile Support for the VS Code Debugger (#889)
We can't access any of the vscode api in the `npm` module, and that
module gets called from within a worker, limiting our ability to pass
functions back and forth. Instead, we can directly access the wasm in
the `vscode` module).

Things that don't work, that I tried:
1. Loading the project in the `InlineDebugAdapterFactory` constructor.
Can't have an async constructor.
2. Loading the files in the `npm` package at all. Can't access
`vscode.fs` related APIs, and can't pass them in as callbacks.
3. Loading the files in the `wasm` layer -- the callbacks have to pass
through a worker boundary via `npm`, so that doesn't fly either.

# Summary

- Refactor the transformers for JS async callbacks, for better
readability
[1](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/wasm/src/language_service.rs#L32)
- Create a shared project loader, to be used by both the debugger and
the language service (eventually)
[1](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/compiler/qsc_project/src/js.rs#L31)
[2](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/wasm/src/project_system.rs)
- Updates the debugger to load all project sources in the `vscode`
layer.
[1](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/vscode/src/debugger/session.ts#L113)
[2](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/vscode/src/debugger/session.ts#L134).

# Notes

We discussed a few ways of approaching this:

One was to pass callbacks in through the JS, all the way into the WASM
-- the way the language service does. This, of course, doesn't work in
workers, so it won't work for the debugger.

The next thought was to directly construct the `vscode.fs` callbacks in
the `npm` directory, instead of passing them in. That way, the callbacks
never cross the worker boundary. Unfortunately, we don't have access to
`vscode.fs` in `npm`, so that would not work either.

The final approach, and the one that I ultimately ended up implementing,
is to provide a more minimal project loading API from the WASM, allowing
the `/vscode` layer to directly access our project system and load the
sources ahead of time, before the worker ever needs them. This allows us
to pass them in to the worker as strings, which are structurally
cloneable types. This works!

The difficult (impossibility) to work with here is that the `/vscode`
crate didn't have direct access to the project logic, which is contained
within the wasm. Because of that, I had to directly access the wasm
types from `vscode`. I'm not sure we want this, but I'm also not sure of
any other way to get access to `qsc_project` while in `/vscode` , which
is the *only* place we can actually use `vscode.fs`.


TODO: 
- [x] docstrings on transformer funcs
- [x] build consensus and validate that importing wasm into `vscode` is
okay
- [x] does this manner of importing wasm even work in a packaged vsix?
[1](https://github.com/microsoft/qsharp/blob/983b875855cb0edfdf4ad0dacf3a0189c84b3df4/vscode/src/debugger/session.ts#L8)
  - [x] Actually, I should wrap it in an API in `npm`
- [x] figure out what the correct thing to do is
[here](https://github.com/microsoft/qsharp/pull/889/files#diff-c2b7e9f81359638edbfc422c7f3df0deda6bd218866e6ebff36e57a85dc780c6R141)
- [x] test in windows

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>
Co-authored-by: Ian Davis <ian.f.davis@microsoft.com>