Use more targeted spans for type equality errors (#903)
This updates a few spots where type equality constraints were used in
the type checker and had the whole expression span rather than just a
component of the expression. This makes it easier to identify specific
parts of an expression with type errors, especially for larger
expressions like addition of multiple variables.


![image](https://github.com/microsoft/qsharp/assets/10567287/a3bfb93a-2a9e-4d02-9c78-ad1e8bcdf52b)