Fix bug in qubit reset on measure logic (#939)
For base profile, we need to "reset" the qubit identifier for qubits
that are measured such that later use of the same ID gets a fresh
hardware ID. Because the underlying bug was a subtle misuse of the two
types (where we have a map from `usize` to `usize`), this change adds a
little bit of extra type safety by ensuring that mapped qubits are a new
wrapper type `HardwareId`. Added a test case that confirms qubit reuse
behaves as expected.

Thanks @cesarzc for figuring this one out!