Prioritize open file buffers over files on the fs in the extension (#925)
I have one primary question about this: is there a way to get the
original source contents for an `OpenDocument` back from a
`Compilation`? I could not find one, and that's why I added
`latest_str_content`, to keep track of the latest `str` reported by
`update_document`. It isn't the most elegant, but it might be the only
option, since `Compilation` doesn't expose the original input at all.

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>