Fix pre-existing iterator bug that was exposed by multi-file (#952)
We had an iterator implementation which worked in single-file mode, but
iterated infinitely in multi-file mode. This was a subtle bug, and we
fixed it by switching to a simpler iterator construct.

We also switched to only instantiating one project loader for the
language service, instead of constructing one-per-load-document,
significantly reducing wasm-heap presssure.