Updated comparisons to make them more useful (#949)
Updated comparison operations.

* Replaced operations that inverted target qubit based on comparison
result of a BigInt and LE register with operations that apply action to
a target. The new operations are more general operations, and they are
needed for the upcoming samples.
* Removed operations that inverted target qubit based on comparison
result of two LE registers. We already have corresponding set of more
general operations that would apply actions.
* Previously all comparisons except for equality comparisons allowed
singly controlled actions. Actions had to support multiqubit control to
be passed into equality comparisons. This restriction has been lifted.
Actions that only support single-qubit control can now be passed to
equality comparisons.

NOTE that the former more specific functions didn't implement any
optimizations based on the fact that the action was an X gate. Also, to
use the old functionality one needs to pass an X gate as the first
argument.

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>
Co-authored-by: Scott Carda <55811729+ScottCarda-MS@users.noreply.github.com>