Fix garbage offsets in completion list (#968)
Repro:

1. Have a folder with `qsharp.json`, `foo.qs` and `bar.qs`
2. In `foo.qs` (name matters because alphabetical order matters for this
repro)

```qsharp
namespace Foo {
}
```

3. Open `foo.qs` which should be an empty file
4. Start typing `n` to bring up completion list, watch output log:

```
2024-01-02 15:21:16.319 [error] Failed to map all utf8 indexes. Remaining indexes are: 21
2024-01-02 15:21:16.320 [error] Failed to map all utf8 indexes. Remaining indexes are: 21
2024-01-02 15:21:16.320 [error] Failed to map all utf8 indexes. Remaining indexes are: 21
2024-01-02 15:21:16.320 [error] Failed to map all utf8 indexes. Remaining indexes are: 21
2024-01-02 15:21:16.320 [error] Failed to map all utf8 indexes. Remaining indexes are: 21
......etc......
```


The problem comes from the fact that `get_completions` gets called
*before* the current document is asynchronously updated.
`get_completions` is called with offset 1, which is an offset that is
out of bounds for the current contents of the document. We end up
resolving the offset to the next document in the project, `foo.qs`. It's
just chaos from there.

To fix this, when we receive an offset that's past the end of the
current document we should just use the document end offset instead.