Diagnostics event for closed document is not really an error (#966)
Just changing the severity of the log since we can actually hit this in
practice.

Repro:
1. Create a new untitled file.
2. Change the language to Q#.
3. Type some bad Q# into the file to make errors show up.
4. Save the file to disk as `foo.qs`.
5. Observe the Output window for an error log.

The reason for this error is that the rename gets handled as a pair of
events: `update_document`, under the new name, followed by
`close_document` under the old name. The first event causes all the
diagnostics to get republished, including the ones associated with the
old name (this is where we log the error). The second event clears the
diagnostics associated with the old name. Then all is good.