[API Proposal] Array CircularlyShifted array function (#881)
I recently needed that and was surprised that in the reach pool of
built-in array functions there was no rotation function (rotate left or
right).

This PR adds a `Rotated` function which performs just that. It follows
the pattern of `Padded` function, in that a positive step indicates
rotation right, and negative indicates a rotation left.

Hope this makes sense.