Fix notebook language service panics (#990)
Fixes #873

ISSUE 1 (#873):

`close_notebook_document` takes both the notebook URI and a list of cell
URIs, which is redundant, since we should already know in the language
service which cells belong to the notebook. We still had an assert to
verify that these two parameters were being sent consistently - and it
turns out they weren't. When the last Q# cell in a notebook is deleted,
we send `close_notebook_document` with _no_ Q# cells (instead of the
last known list). That tripped the assert.

I'm just removing the redundant `cell_uris` parameter and using the open
document map instead.

ISSUE 2 (new panic):

I'm 99.99% sure this is a change in VS Code, since this panic is so
trivial to hit. All you have to do is close a notebook. I tried with
builds of `qsharp` from September, and I still hit it. Basically, using
`vscode.languages.match()` to determine whether a notebook cell is Q# is
very unreliable, as it returns no match if the notebook is being closed.
This leads to a panic in the language service. So instead I'm just using
the `languageId` and `scheme` - same logic, much simpler.